Only update state file if not already up to date.
Case 89625: In some cases, the state file was being unnecessarily
rewritten every time a cPanel::TaskQueue object was constructed,
just to ensure that it had an up to date copy of certain paramters
passed into the constructor. Apart from being wasteful, this also
sometimes had destructive consequences if several TaskQueue objects
were attempting to update it in parallel.

Fix this by only rewriting the file if the values have actually
changed relative to what was already in the file.